Fixing a possible deadlock in NEW-ORDER.
We are selecting the stock rows for update. There is
a chance that another concurrent transaction is ordering
the same items from the same warehouses but that the order
lines are in a different sort order. To avoid the resulting
deadlock we must sort the order lines by itemID.

(cherry picked from commit d94d2c206669bed4b44ff0ce373e71f1e4fbe4c8)