Make result data collection scale
The previous way of writing a CSV result line per
transaction did not scale. The system would collect
several GB per hour for large DB servers under test.
Post processing that with R is not reasonable.

We now write a summary CSV line per Monkey thread
at a configurable interval (default 10 seconds). This
only has the transaction type, number and sums in
milliseconds for delay and latency. This is sufficient
to create the performance graphs. We also write out
the histogram data so that a report generator can
determine latency percentiles from that. We might
want to make the number of edges (1000) and the cutoff
value (600 seconds) config parameters.

This commit completely breaks the HTML report post
processing code. The plan is to replace that code with
Python scripts that are numpy and matplotlib based
anyway.