Add support for Babelfish
Babelfish for PostgreSQL is an Amazon project aiming to
add MSSQL-Server compatibility to PostgreSQL. The project
is currently in the early stages and misses critical
features. One of them is the lack of the ANSI-SQL support
of 'SELECT ... FOR UPDATE'. T-SQL has that functionality
but it is under a query HINT. Babelfish currently ignores
all hints. This commit is a workaround for this problem
and needs to be revised when Babelfish matures.