Add config parameters to run only on a range of warehouses
Add config parameters useWarehouseFrom and useWarehouseTo
to be able to perform a benchmark run on only a range of
the total configured warehouses.

This is intended to run BenchmarkSQL on a distributed
database setup and separate the warehouses onto different
"home nodes". In such a setup only the CUSTOMER and STOCK
tables will experience update-update conflicts on the
PAYMENT and NEW_ORDER transactions respectively.